Split postgres_controller into separate parts for each engine
