+++ internal/controller/postgres_controller.go
@@ -11,11 +11,7 @@ import (
 	cnpgv1 "github.com/cloudnative-pg/cloudnative-pg/api/v1"
 	"github.com/nais/pgrator/internal/config"
 	"github.com/nais/pgrator/internal/namegen"
-	rccnpg "github.com/nais/pgrator/internal/resourcecreator/cnpg"
 	rciam "github.com/nais/pgrator/internal/resourcecreator/iam"
-	rcmonitoring "github.com/nais/pgrator/internal/resourcecreator/monitoring"
-	rcnetpol "github.com/nais/pgrator/internal/resourcecreator/netpol"
-	rczalando "github.com/nais/pgrator/internal/resourcecreator/zalando"
 	"github.com/nais/pgrator/internal/synchronizer/action"
 	"github.com/nais/pgrator/internal/synchronizer/events"
 	"github.com/nais/pgrator/internal/synchronizer/reconciler"
@@ -179,95 +175,6 @@ func (r *PostgresReconciler) Update(obj *data_nais_io_v1.Postgres, preparedData 
 	}
 }
 
-func (r *PostgresReconciler) updateCNPG(obj *data_nais_io_v1.Postgres, preparedData PreparedData, relatedObjects reconciler.RelatedObjects) ([]action.Action, ctrl.Result, error) {
-	pgClusterName, pgNamespace, err := getClusterNameAndNamespace(obj, api.EngineCNPG)
-	if err != nil {
-		return nil, ctrl.Result{}, err
-	}
-
-	var actions []action.Action
-
-	cluster, err := rccnpg.CreateClusterSpec(obj, r.Config, pgClusterName, pgNamespace)
-	if err != nil {
-		return nil, ctrl.Result{}, err
-	}
-	existingCluster := relatedObjects.GetMatching(cluster)
-	if existingCluster != nil {
-		actions = append(actions, action.Update(cluster, obj, cnpgClusterConditionGetter, r.Recorder))
-	} else {
-		actions = append(actions, action.Create(cluster, obj, cnpgClusterConditionGetter, r.Recorder))
-	}
-
-	if r.Config.CNPG.BackupBucket != "" {
-		backup := rccnpg.CreateScheduledBackup(obj, r.Config, pgClusterName, pgNamespace)
-		actions = append(actions, action.CreateOrUpdate(backup, obj, existsConditionGetter, r.Recorder))
-	}
-
-	pooler := rccnpg.CreatePooler(obj, pgClusterName, pgNamespace)
-	existingPooler := relatedObjects.GetMatching(pooler)
-	if existingPooler != nil {
-		actions = append(actions, action.Update(pooler, obj, existsConditionGetter, r.Recorder))
-	} else {
-		actions = append(actions, action.Create(pooler, obj, existsConditionGetter, r.Recorder))
-	}
-
-	cnpgNetpol := rcnetpol.CreateCNPG(obj, pgClusterName, pgNamespace)
-	actions = append(actions, action.CreateOrUpdate(cnpgNetpol, obj, existsConditionGetter, r.Recorder))
-
-	// TODO: Set up IAM
-	iamActions, err := r.cnpgIAMActions(obj, preparedData, pgNamespace, r.Config.CNPG.BackupBucket, relatedObjects)
-	if err != nil {
-		return nil, ctrl.Result{}, err
-	}
-	actions = append(actions, iamActions...)
-
-	if !r.Config.PrometheusRulesDisabled {
-		prometheusRule := rcmonitoring.CreateCNPGPrometheusRule(obj, pgClusterName, pgNamespace)
-		actions = append(actions, action.CreateOrUpdate(prometheusRule, obj, existsConditionGetter, r.Recorder))
-
-		podMonitor := rcmonitoring.CreateCNPGPodMonitor(obj, pgClusterName, pgNamespace)
-		actions = append(actions, action.CreateOrUpdate(podMonitor, obj, existsConditionGetter, r.Recorder))
-	}
-
-	return actions, ctrl.Result{}, nil
-}
-
-func (r *PostgresReconciler) updateZalando(obj *data_nais_io_v1.Postgres, preparedData PreparedData, relatedObjects reconciler.RelatedObjects) ([]action.Action, ctrl.Result, error) {
-	pgClusterName, pgNamespace, err := getClusterNameAndNamespace(obj, api.EngineZalando)
-	if err != nil {
-		return nil, ctrl.Result{}, err
-	}
-
-	var actions []action.Action
-	cluster, err := rczalando.CreateClusterSpec(obj, r.Config, pgClusterName, pgNamespace)
-	if err != nil {
-		return nil, ctrl.Result{}, err
-	}
-
-	existingCluster := relatedObjects.GetMatching(cluster)
-	if existingCluster != nil {
-		actions = append(actions, action.Update(cluster, obj, postgresqlConditionGetter, r.Recorder))
-	} else {
-		actions = append(actions, action.Create(cluster, obj, postgresqlConditionGetter, r.Recorder))
-	}
-
-	zalandoNetpol := rcnetpol.CreateZalando(obj, pgClusterName, pgNamespace)
-	actions = append(actions, action.CreateOrUpdate(zalandoNetpol, obj, existsConditionGetter, r.Recorder))
-
-	iamActions, err := r.zalandoIAMActions(obj, preparedData, pgNamespace, r.Config.WalGsBucket, relatedObjects)
-	if err != nil {
-		return nil, ctrl.Result{}, err
-	}
-	actions = append(actions, iamActions...)
-
-	if !r.Config.PrometheusRulesDisabled {
-		prometheusRule := rcmonitoring.CreateZalandoPrometheusRule(obj, pgClusterName, pgNamespace)
-		actions = append(actions, action.CreateOrUpdate(prometheusRule, obj, existsConditionGetter, r.Recorder))
-	}
-
-	return actions, ctrl.Result{}, nil
-}
-
 func iamConditionGetter(obj client.Object, scheme *runtime.Scheme) []meta_v1.Condition {
 	var iamConditions []meta_v1.Condition
 	switch o := obj.(type) {
@@ -319,42 +226,6 @@ func iamConditionGetter(obj client.Object, scheme *runtime.Scheme) []meta_v1.Con
 	return result
 }
 
-func postgresqlConditionGetter(obj client.Object, scheme *runtime.Scheme) []meta_v1.Condition {
-	pg := obj.(*acid_zalan_do_v1.Postgresql)
-
-	conditions := []conditionConfig{
-		{
-			Type:   "Available",
-			Status: pg.Status.PostgresClusterStatus == acid_zalan_do_v1.ClusterStatusRunning || pg.Status.PostgresClusterStatus == acid_zalan_do_v1.ClusterStatusUpdating,
-		},
-		{
-			Type:   "Progressing",
-			Status: pg.Status.PostgresClusterStatus == acid_zalan_do_v1.ClusterStatusCreating || pg.Status.PostgresClusterStatus == acid_zalan_do_v1.ClusterStatusUpdating,
-		},
-		{
-			Type:   "Degraded",
-			Status: !pg.Status.Success(),
-		},
-	}
-
-	statusReason := pg.Status.String()
-	if statusReason == "" {
-		statusReason = "Unknown"
-	}
-
-	result := make([]meta_v1.Condition, 0, len(conditions))
-	for _, condition := range conditions {
-		result = append(result, meta_v1.Condition{
-			Type:               fmt.Sprintf("%s/%s", typePrefix(obj, scheme), condition.Type),
-			Status:             makeCondition(condition.Status),
-			ObservedGeneration: obj.GetGeneration(),
-			Reason:             statusReason,
-		})
-	}
-
-	return result
-}
-
 func (r *PostgresReconciler) Delete(obj *data_nais_io_v1.Postgres, preparedData PreparedData, relatedObjects reconciler.RelatedObjects) ([]action.Action, ctrl.Result, error) {
 	switch preparedData.Engine {
 	case api.EngineCNPG:
@@ -362,79 +233,6 @@ func (r *PostgresReconciler) Delete(obj *data_nais_io_v1.Postgres, preparedData 
 	default:
 		return r.deleteZalando(obj, preparedData, relatedObjects)
 	}
-}
-
-func (r *PostgresReconciler) deleteCNPG(obj *data_nais_io_v1.Postgres, preparedData PreparedData, relatedObjects reconciler.RelatedObjects) ([]action.Action, ctrl.Result, error) {
-	actionFunc := action.DeleteIfExists
-	sharedActionFunc := action.Unclaim
-	if !obj.Spec.Cluster.AllowDeletion {
-		actionFunc = action.NoOp
-		sharedActionFunc = action.NoOp
-	}
-
-	pgClusterName, pgNamespace, err := getClusterNameAndNamespace(obj, api.EngineCNPG)
-	if err != nil {
-		return nil, ctrl.Result{}, err
-	}
-
-	actions := make([]action.Action, 0, 4)
-
-	cluster := rccnpg.MinimalCluster(obj, pgClusterName, pgNamespace)
-	actions = append(actions, actionFunc(cluster, obj, cnpgClusterConditionGetter, r.Recorder))
-
-	backup := rccnpg.MinimalScheduledBackup(obj, pgClusterName, pgNamespace)
-	actions = append(actions, actionFunc(backup, obj, existsConditionGetter, r.Recorder))
-
-	pooler := rccnpg.MinimalPooler(obj, pgClusterName, pgNamespace)
-	actions = append(actions, actionFunc(pooler, obj, existsConditionGetter, r.Recorder))
-
-	cnpgNetpol := rcnetpol.Minimal(obj, pgClusterName, pgNamespace)
-	actions = append(actions, actionFunc(cnpgNetpol, obj, existsConditionGetter, r.Recorder))
-
-	iamActions := r.deleteIAMActions(obj, preparedData, pgNamespace, r.Config.CNPG.BackupBucket, sharedActionFunc, relatedObjects)
-	actions = append(actions, iamActions...)
-
-	if !r.Config.PrometheusRulesDisabled {
-		prometheusRule := rcmonitoring.MinimalPrometheusRule(obj, pgClusterName)
-		actions = append(actions, actionFunc(prometheusRule, obj, existsConditionGetter, r.Recorder))
-
-		podMonitor := rcmonitoring.MinimalCNPGPodMonitor(obj, pgClusterName, pgNamespace)
-		actions = append(actions, actionFunc(podMonitor, obj, existsConditionGetter, r.Recorder))
-	}
-
-	return actions, ctrl.Result{}, nil
-}
-
-func (r *PostgresReconciler) deleteZalando(obj *data_nais_io_v1.Postgres, preparedData PreparedData, relatedObjects reconciler.RelatedObjects) ([]action.Action, ctrl.Result, error) {
-	actionFunc := action.DeleteIfExists
-	sharedActionFunc := action.Unclaim
-	if !obj.Spec.Cluster.AllowDeletion {
-		actionFunc = action.NoOp
-		sharedActionFunc = action.NoOp
-	}
-
-	pgClusterName, pgNamespace, err := getClusterNameAndNamespace(obj, api.EngineZalando)
-	if err != nil {
-		return nil, ctrl.Result{}, err
-	}
-
-	var actions []action.Action
-
-	cluster := rczalando.MinimalCluster(obj, pgClusterName, pgNamespace)
-	actions = append(actions, actionFunc(cluster, obj, postgresqlConditionGetter, r.Recorder))
-
-	zalandoNetpol := rcnetpol.Minimal(obj, pgClusterName, pgNamespace)
-	actions = append(actions, actionFunc(zalandoNetpol, obj, existsConditionGetter, r.Recorder))
-
-	iamActions := r.deleteIAMActions(obj, preparedData, pgNamespace, r.Config.WalGsBucket, sharedActionFunc, relatedObjects)
-	actions = append(actions, iamActions...)
-
-	if !r.Config.PrometheusRulesDisabled {
-		prometheusRule := rcmonitoring.MinimalPrometheusRule(obj, pgClusterName)
-		actions = append(actions, actionFunc(prometheusRule, obj, existsConditionGetter, r.Recorder))
-	}
-
-	return actions, ctrl.Result{}, nil
 }
 
 type actionFunc func(client.Object, api.NaisObject, action.ConditionGetter, events.Recorder) action.Action
@@ -588,143 +386,6 @@ func validateVersionForEngine(majorVersion string, engine string) error {
 		}
 	}
 	return nil
-}
-
-// zalandoIAMActions returns the IAM actions used by Zalando
-func (r *PostgresReconciler) zalandoIAMActions(obj *data_nais_io_v1.Postgres, preparedData PreparedData, pgNamespace, backupBucket string, relatedObjects reconciler.RelatedObjects) ([]action.Action, error) {
-	var actions []action.Action
-
-	workloadIdentityPolicyName, storageBucketPolicyName, logsWriterPolicyName := IAMPolicyMemberNames(obj.GetNamespace())
-
-	workloadIdentityPolicy := rciam.CreateWorkloadIdentityPolicyMember(workloadIdentityPolicyName, obj.GetNamespace(), pgNamespace, r.Config.GoogleProjectID, GSAName, KSAName)
-	existingWorkloadIdentityPolicy := relatedObjects.GetMatching(workloadIdentityPolicy)
-	if existingWorkloadIdentityPolicy == nil {
-		actions = append(actions, action.Create(workloadIdentityPolicy, obj, iamConditionGetter, r.Recorder))
-	} else if iamPolicyHasChanges(workloadIdentityPolicy, existingWorkloadIdentityPolicy.(*iam_cnrm_cloud_google_com_v1beta1.IAMPolicyMember)) {
-		if r.Config.ResyncIAMPermissions {
-			actions = append(actions, action.Recreate(workloadIdentityPolicy, obj, iamConditionGetter, r.Recorder))
-		} else {
-			return nil, fmt.Errorf("want to change IAMPolicyMember %s, but configuration does not allow recreate", client.ObjectKeyFromObject(workloadIdentityPolicy))
-		}
-	} else {
-		actions = append(actions, action.Claim(workloadIdentityPolicy, obj, iamConditionGetter, r.Recorder))
-	}
-
-	if backupBucket != "" {
-		storageBucketPolicy := rciam.CreateStorageBucketPolicyMember(storageBucketPolicyName, ServiceAccountsNamespace, preparedData.TeamGoogleProjectID, GSAName, backupBucket)
-		existingStorageBucketPolicy := relatedObjects.GetMatching(storageBucketPolicy)
-		if existingStorageBucketPolicy == nil {
-			actions = append(actions, action.Create(storageBucketPolicy, obj, iamConditionGetter, r.Recorder))
-		} else if iamPolicyHasChanges(storageBucketPolicy, existingStorageBucketPolicy.(*iam_cnrm_cloud_google_com_v1beta1.IAMPolicyMember)) {
-			if r.Config.ResyncIAMPermissions {
-				actions = append(actions, action.Recreate(storageBucketPolicy, obj, iamConditionGetter, r.Recorder))
-			} else {
-				return nil, fmt.Errorf("want to change IAMPolicyMember %s, but configuration does not allow recreate", client.ObjectKeyFromObject(storageBucketPolicy))
-			}
-		} else {
-			actions = append(actions, action.Claim(storageBucketPolicy, obj, iamConditionGetter, r.Recorder))
-		}
-	}
-
-	logsWriterPolicy := rciam.CreateLogsWriterPolicyMember(logsWriterPolicyName, obj.GetNamespace(), preparedData.TeamGoogleProjectID, GSAName)
-	existingLogsWriterPolicy := relatedObjects.GetMatching(logsWriterPolicy)
-	if existingLogsWriterPolicy == nil {
-		actions = append(actions, action.Create(logsWriterPolicy, obj, iamConditionGetter, r.Recorder))
-	} else if iamPolicyHasChanges(logsWriterPolicy, existingLogsWriterPolicy.(*iam_cnrm_cloud_google_com_v1beta1.IAMPolicyMember)) {
-		if r.Config.ResyncIAMPermissions {
-			actions = append(actions, action.Recreate(logsWriterPolicy, obj, iamConditionGetter, r.Recorder))
-		} else {
-			return nil, fmt.Errorf("want to change IAMPolicyMember %s, but configuration does not allow recreate", client.ObjectKeyFromObject(logsWriterPolicy))
-		}
-	} else {
-		actions = append(actions, action.Claim(logsWriterPolicy, obj, iamConditionGetter, r.Recorder))
-	}
-
-	gsa := rciam.CreateIAMServiceAccount(GSAName, obj.GetNamespace())
-	existingGsa := relatedObjects.GetMatching(gsa)
-	if existingGsa != nil {
-		actions = append(actions, action.Claim(gsa, obj, iamConditionGetter, r.Recorder))
-	} else {
-		actions = append(actions, action.Create(gsa, obj, iamConditionGetter, r.Recorder))
-	}
-
-	kubernetesSA := rciam.CreateKubernetesServiceAccount(KSAName, pgNamespace, preparedData.TeamGoogleProjectID, GSAName)
-	existingKubernetesSA := relatedObjects.GetMatching(kubernetesSA)
-	if existingKubernetesSA != nil {
-		actions = append(actions, action.Update(kubernetesSA, obj, existsConditionGetter, r.Recorder))
-	} else {
-		actions = append(actions, action.Create(kubernetesSA, obj, existsConditionGetter, r.Recorder))
-	}
-
-	postgresPodRoleBinding := rciam.CreateRoleBinding(RoleBindingName, KSAName, ClusterRoleName, pgNamespace)
-	existingPRB := relatedObjects.GetMatching(postgresPodRoleBinding)
-	if existingPRB != nil {
-		actions = append(actions, action.Update(postgresPodRoleBinding, obj, existsConditionGetter, r.Recorder))
-	} else {
-		actions = append(actions, action.Create(postgresPodRoleBinding, obj, existsConditionGetter, r.Recorder))
-	}
-
-	return actions, nil
-}
-
-// cnpgIAMActions returns the IAM actions used by CNPG
-func (r *PostgresReconciler) cnpgIAMActions(obj *data_nais_io_v1.Postgres, preparedData PreparedData, pgNamespace, backupBucket string, relatedObjects reconciler.RelatedObjects) ([]action.Action, error) {
-	// TODO: Implement IAM
-	return nil, nil
-}
-
-func cnpgClusterConditionGetter(obj client.Object, scheme *runtime.Scheme) []meta_v1.Condition {
-	cluster := obj.(*cnpgv1.Cluster)
-
-	isReady := false
-	for _, cond := range cluster.Status.Conditions {
-		if cond.Type == "Ready" && cond.Status == meta_v1.ConditionTrue {
-			isReady = true
-			break
-		}
-	}
-
-	conditions := []conditionConfig{
-		{
-			Type:   "Available",
-			Status: isReady,
-		},
-		{
-			Type: "Progressing",
-			Status: cluster.Status.Phase == cnpgv1.PhaseFirstPrimary ||
-				cluster.Status.Phase == cnpgv1.PhaseCreatingReplica ||
-				cluster.Status.Phase == cnpgv1.PhaseUpgrade ||
-				cluster.Status.Phase == cnpgv1.PhaseOnlineUpgrading ||
-				cluster.Status.Phase == cnpgv1.PhaseApplyingConfiguration ||
-				cluster.Status.Phase == cnpgv1.PhaseSwitchover,
-		},
-		{
-			Type: "Degraded",
-			Status: cluster.Status.Phase == cnpgv1.PhaseUnrecoverable ||
-				cluster.Status.Phase == cnpgv1.PhaseImageCatalogError ||
-				cluster.Status.ReadyInstances < cluster.Status.Instances,
-		},
-	}
-
-	message := cluster.Status.Phase
-	if message == "" {
-		message = "Unknown"
-	}
-
-	reason := reasonable(message)
-
-	result := make([]meta_v1.Condition, 0, len(conditions))
-	for _, condition := range conditions {
-		result = append(result, meta_v1.Condition{
-			Type:               fmt.Sprintf("%s/%s", typePrefix(obj, scheme), condition.Type),
-			Status:             makeCondition(condition.Status),
-			ObservedGeneration: obj.GetGeneration(),
-			Reason:             reason,
-			Message:            message,
-		})
-	}
-
-	return result
 }
 
 func reasonable(message string) string {
